requirements.txt
python-telegram-bot==21.10
aiohttp==3.11.11
python-dotenv==1.0.1
